Add cache for RestaurantService and second level hibernate

+++ src/test/java/ru/imakabr/votingsystem/service/VoteServiceTest.java
@@ -3,13 +3,12 @@ package ru.imakabr.votingsystem.service;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
-import ru.imakabr.votingsystem.RestaurantTestData;
 import ru.imakabr.votingsystem.UserTestData;
 import ru.imakabr.votingsystem.VoteTestData;
-import ru.imakabr.votingsystem.model.Restaurant;
 import ru.imakabr.votingsystem.model.User;
 import ru.imakabr.votingsystem.model.Vote;
 import ru.imakabr.votingsystem.util.exception.NotFoundException;
+import ru.imakabr.votingsystem.util.exception.TimeVoteLimitException;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -26,26 +25,41 @@ public class VoteServiceTest extends AbstractServiceTest {
 
     @Test
     void create() throws Exception {
-        Vote newVote = new Vote(null, UserTestData.USER, TOKYO_CITY, LocalDate.now());
-        Vote created = voteService.create(TOKYO_CITY, USER_ID);
-        newVote.setId(created.getId());
-        VoteTestData.assertMatch(created, newVote);
+        if (checkTime()) {
+            Vote newVote = new Vote(null, UserTestData.USER, TOKYO_CITY, LocalDate.now());
+            Vote created = voteService.create(TOKYO_CITY, USER_ID);
+            newVote.setId(created.getId());
+            VoteTestData.assertMatch(created, newVote);
+        } else {
+            assertThrows(TimeVoteLimitException.class, () ->
+                    voteService.create(TOKYO_CITY, USER_ID));
+        }
     }
 
     @Test
     void duplicateCreate() throws Exception {
-        voteService.create(TOKYO_CITY, USER_ID);
-        assertThrows(DataAccessException.class, () ->
-                voteService.create(KWAKINN, USER_ID));
+        if (checkTime()) {
+            voteService.create(TOKYO_CITY, USER_ID);
+            assertThrows(DataAccessException.class, () ->
+                    voteService.create(KWAKINN, USER_ID));
+        } else {
+            assertThrows(TimeVoteLimitException.class, () ->
+                    voteService.create(TOKYO_CITY, USER_ID));
+        }
     }
 
     @Test
     void delete() throws Exception {
-        voteService.create(TOKYO_CITY, USER_ID);
-        voteService.delete(VOTE_ID+3, USER_ID);
-        List<Vote> votes = voteService.getAllVotesByUserId(USER_ID);
-        assertMatch(votes.get(0).getRestaurant(), KETCH_UP);
-        assertMatch(votes.get(1).getRestaurant(), TOKYO_CITY);
+        if (checkTime()) {
+            voteService.create(TOKYO_CITY, USER_ID);
+            voteService.delete(VOTE_ID + 3, USER_ID);
+            List<Vote> votes = voteService.getAllVotesByUserId(USER_ID);
+            assertMatch(votes.get(0).getRestaurant(), KETCH_UP);
+            assertMatch(votes.get(1).getRestaurant(), TOKYO_CITY);
+        } else {
+            assertThrows(TimeVoteLimitException.class, () ->
+                    voteService.create(TOKYO_CITY, USER_ID));
+        }
     }
 
     @Test
@@ -65,13 +79,6 @@ public class VoteServiceTest extends AbstractServiceTest {
         assertThrows(NotFoundException.class, () ->
                 voteService.get(1, USER_ID));
     }
-
-//    @Test
-//    void update() throws Exception {
-//        Vote updated = new Vote(VOTE_ID, UserTestData.USER, KWAKINN, LocalDate.of(2019, 9, 20));
-//        voteService.update(TOKYO_CITY, USER_ID, VOTE_ID);
-//        assertMatch(voteService.getRestaurantByUserIdAndDate(USER_ID, LocalDate.of(2019, 9, 20)), KWAKINN);
-//    }
 
     @Test
     void getAllVotesByUserId() {
